Move `test` into `tests` folder
